DB関連ファイルをフォルダ別で管理
