+++ discordbot.py
@@ -1,18 +1,10 @@
 import discord
 from discord import app_commands
-import os
-from dotenv import load_dotenv
+from apikey import TOKEN,channelid
 
 """
 discordbot初期設定
 """
-#bot Tokenを環境ファイルから読み込み
-load_dotenv()
-TOKEN=os.environ["TOKEN"]
-
-#channelid
-channelid="entry channelid"
-
 #初期設定
 intents=discord.Intents.default()
 intents.messages=True
