+++ processed_id_list.csv
@@ -1,4 +1,70 @@
 0
+5126659
+5126960
+5127008
+5126973
+5126887
+5127123
+5122745
+5127092
+5127107
+5127221
+5127198
+5127274
+5127298
+5126925
+5119473
+5127370
+5127401
+4825038
+5127490
+5127545
+5127565
+5127569
+5127597
+5127626
+5127644
+5127710
+5122647
+5127737
+5127735
+5127766
+5118532
+5127792
+5127866
+5126075
+5127915
+5057416
+5100731
+5124880
+5082809
+5099562
+5118378
+5103230
+5127963
+5018540
+5050845
+5082884
+5103033
+5125870
+5116092
+5127241
+5117185
+5128035
+5127703
+5128084
+5127996
+5128316
+5128359
+5128386
+5128293
+5128422
+5083440
+5128419
+5128213
+5128483
+5128522
+5076511
 5088385
 5126964
 5036271
@@ -833,69 +899,3 @@ item_4995093
 item_4994540
 item_4989849
 item_4991241
-item_4994356
-item_4994754
-item_4965500
-item_4994455
-item_4990465
-item_4461558
-item_4990572
-item_4990302
-item_4990486
-item_4990646
-item_4990643
-item_4990758
-item_4990817
-item_4990805
-item_4990799
-item_4989389
-item_4990952
-item_4990955
-item_4990819
-item_4991008
-item_4991019
-item_4991017
-item_4991039
-item_4991062
-item_4988536
-item_4991077
-item_4991113
-item_4991100
-item_4987499
-item_4991128
-item_4991119
-item_4915091
-item_4991174
-item_4991204
-item_4991254
-item_4991287
-item_4991306
-item_4991328
-item_4991314
-item_4988183
-item_4991366
-item_4991312
-item_4984730
-item_4989584
-item_4991088
-item_4991543
-item_4991675
-item_4991650
-item_4991687
-item_4991785
-item_4991765
-item_4991898
-item_4991271
-item_4992021
-item_4992082
-item_4991196
-item_4992141
-item_4987980
-item_4992312
-item_4992284
-item_4964116
-item_4992380
-item_4992325
-item_4992398
-item_4990927
-item_4992664
